+++ druid/src/text/mod.rs
@@ -7,14 +7,7 @@ mod attribute;
 mod backspace;
 mod editable_text;
 mod font_descriptor;
-
-#[deprecated(since = "0.8.0", note = "use types from druid::text module instead")]
-#[doc(hidden)]
-pub mod format;
-// a hack to let us deprecate the format module; we can remove this when we make
-// format private
-#[path = "format.rs"]
-mod format_priv;
+mod format;
 mod input_component;
 mod input_methods;
 mod layout;
@@ -32,7 +25,7 @@ pub use self::attribute::{Attribute, AttributeSpans, Link};
 pub use self::backspace::offset_for_delete_backwards;
 pub use self::editable_text::{EditableText, EditableTextCursor, StringCursor};
 pub use self::font_descriptor::FontDescriptor;
-pub use self::format_priv::{Formatter, ParseFormatter, Validation, ValidationError};
+pub use self::format::{Formatter, ParseFormatter, Validation, ValidationError};
 pub use self::layout::{LayoutMetrics, TextLayout};
 pub use self::movement::movement;
 pub use input_component::{EditSession, TextComponent};